refactor schema tests into separate modules
